Refactored ALL of the IIIF code
