+++ src/components/Thumbnail/Thumbnail.tsx
@@ -1,8 +1,8 @@
 import { useInView } from 'react-intersection-observer';
 import { CanvasInformation, FileImage, LoadedImage } from '@/model';
 import { useImages } from '@/store';
-import { getThumbnail } from '@/utils/iiif/lib/helpers';
 import { Spinner } from '../Spinner';
+import { getThumbnailURL } from '@/utils/cozy-iiif';
 
 interface ThumbnailProps {
 
@@ -23,7 +23,7 @@ const ThumbnailImage = (props: ThumbnailProps) => {
   const url = loaded && (
     'data' in loaded 
       ? URL.createObjectURL(loaded.data) 
-      : getThumbnail(loaded.canvas, { size: 112 }));
+      : getThumbnailURL(loaded.canvas, 112));
 
   return loaded ? (
     <img
