+++ src/utils/cozy-iiif/index.ts
@@ -1,3 +1,3 @@
-export * from './utils';
-export * from './CozyIIIF';
+export * from './utils/cozy';
+export * from './Cozy';
 export * from './Types';
