+++ src/utils/cozy-iiif/utils/cozy/index.ts
@@ -1,0 +1,1 @@
+export * from './canvas';
