+++ src/utils/cozy-iiif/utils/iiif/index.ts
@@ -1,0 +1,3 @@
+export * from './canvas';
+export * from './imageService';
+export * from './resource';
