+++ src/utils/cozy-iiif/utils/iiif/resource.ts
