--- src/utils/cozy-iiif/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './canvasUtils';
-export * from './imageServiceUtils';
-export * from './resourceUtils';
