--- src/utils/iiif/lib/helpers/index.ts
@@ -1,4 +1,0 @@
-export * from './canvasHelpers';
-export * from './imageHelpers';
-export * from './jsonldHelpers';
-export * from './thumbnailHelpers';
