--- src/utils/iiif/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './identify';
-export * from './parse';
