+++ src/utils/iiif/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './parseIIIFId';
+export * from './getOSDTilesets';
